Box: Add bounding planes (#174)
These will be used in place of a `floor_point` in lacecore. The name `floor_point` makes an assumption about being in a y-up coordinate system, so it's better to let the caller use `mesh.bounding_box.min_y_plane`, or `mesh.bounding_box.min_z_plane` for a z-up coordinate system.