Fix file not found error when deserializing
While this worked in the tests, it did not work after installing the
package because the schema.json was not being shipped.

Rather than complicate the file layout, I've vendored a copy of the
schema into an appropriate place in the Python project, and included a
unit test to ensure the file is kept up to date.